Revert "Rename MigrationBase in VersionedMigration"
This reverts commit feb71f5d14b88a394f85b3b46355af960409cb7a.
