+++ src/Evolve/Migration/MigrationScript.cs
@@ -6,7 +6,7 @@ using Evolve.Utilities;
 
 namespace Evolve.Migration
 {
-    public abstract class MigrationScript : VersionedMigration
+    public abstract class MigrationScript : MigrationBase
     {
         private const string IncorrectMigrationChecksum = "Validate failed: invalid checksum for migration: {0}.";
